osctld: let containers freely configure oom_score_adj
pty-wrapper is run with score -1000 to ensure it will not be killed,
taking down the entire container. The container's init process is
started with oom_score_adj=0, but oom_score_adj_min is -1000, so
services can optionally decrease their score or disable the OOM killer
entirely.

What makes this approach useless is that only user with resource
capability in the initial user namespace can change oom_score_adj_min.
osctld runs the wrapper with -1000, which updates oom_score_adj_min, but
when the wrapper sets the adj to 0, oom_score_adj_min is not changed,
because the wrapper does not have the resource capability. All the
container's processes then inherit the oom_score_adj_min value, meaning
all users inside the container can set oom_score_adj all the way down to
-1000, i.e. to disable the OOM killer for arbitrary processes.